requirements.txt
attrs==19.3.0
aws-cdk.assets==1.28.0
aws-cdk.aws-cloudwatch==1.28.0
aws-cdk.aws-ec2==1.28.0
aws-cdk.aws-efs==1.28.0
aws-cdk.aws-events==1.28.0
aws-cdk.aws-iam==1.28.0
aws-cdk.aws-kms==1.28.0
aws-cdk.aws-lambda==1.28.0
aws-cdk.aws-logs==1.28.0
aws-cdk.aws-s3==1.28.0
aws-cdk.aws-s3-assets==1.28.0
aws-cdk.aws-sqs==1.28.0
aws-cdk.aws-ssm==1.28.0
aws-cdk.core==1.28.0
aws-cdk.cx-api==1.28.0
aws-cdk.region-info==1.28.0
jmespath==0.9.5
jsii==0.22.0

